consolidate preview_kind utilities in preview_mode.ex
