+++ lib/oli_web/live_session_plugs/set_preview_mode.ex
@@ -3,7 +3,7 @@ defmodule OliWeb.LiveSessionPlugs.SetPreviewMode do
 
   import Phoenix.Component, only: [assign: 2]
 
-  alias OliWeb.Delivery.Student.Utils, as: StudentUtils
+  alias OliWeb.Delivery.Instructor.PreviewMode
 
   def on_mount(:default, params, session, socket) do
     preview_mode = socket.assigns[:live_action] == :preview
@@ -15,7 +15,7 @@ defmodule OliWeb.LiveSessionPlugs.SetPreviewMode do
     {:cont,
      assign(socket,
        preview_mode: preview_mode,
-       section_preview_kind: StudentUtils.section_preview_kind(preview_mode, params),
+       section_preview_kind: PreviewMode.section_preview_kind(preview_mode, params),
        template_preview_mode: template_preview_mode,
        template_preview_exit_path:
          if(template_preview_mode, do: ~p"/authoring/template_preview/exit", else: nil)
